Aufgabe 8.5: Projektarbeit

+++ src/main/java/com/team11/Parkhaus/CarIF.java
@@ -3,8 +3,8 @@ package com.team11.Parkhaus;
 public interface CarIF {
     void leave(int duration, int price);
     String getCarType();
-    int getPrice();
-    int getDuration();
+    float getPrice();
+    float getDuration();
     String getTicketId();
     String getLicencePlate();
     String getColor();
